fact-generator/requirements.txt
flask==3.1.0
litellm>=1.70.0
opentelemetry-distro==0.50b0
opentelemetry-exporter-otlp-proto-http==1.29.0
opentelemetry-instrumentation-flask==0.50b0
opentelemetry-instrumentation-httpx==0.50b0
opentelemetry-instrumentation-requests==0.50b0
opentelemetry-sdk==1.29.0
